[#9] Refactor: PayrollItem 컴포넌트 분리
