+++ src/pages/salary-management/Payslip/Payslip.jsx
@@ -1,5 +1,6 @@
 import styled from 'styled-components'
 import { colors } from '@styles/Colors'
+import PayrollItem from './PayrollItem'
 
 export default function Payslip() {
   const payrollData = {
@@ -11,13 +12,6 @@ export default function Payslip() {
     bonus: 200000,
     totalPay: 3800000,
   }
-
-  const PayrollItem = ({ label, value }) => (
-    <PayrollItemContainer>
-      <PayrollLabel>{label}</PayrollLabel>
-      <PayrollValue>{value}</PayrollValue>
-    </PayrollItemContainer>
-  )
 
   return (
     <PayslipContainer>
@@ -80,30 +74,6 @@ const DetailContainer = styled.div`
   flex-direction: column;
   justify-content: space-between;
   margin-bottom: 77px;
-`
-
-const PayrollItemContainer = styled.div`
-  font-size: 16px;
-  display: flex;
-  height: 100%;
-  align-items: center;
-  justify-content: space-between;
-  padding: 8px 40px;
-  border-bottom: 1px solid #e0e0e0;
-
-  &:hover {
-    background-color: #f8fffe;
-    text-decoration: underline;
-  }
-`
-
-const PayrollLabel = styled.div`
-  font-size: 16px;
-  font-weight: 600;
-`
-
-const PayrollValue = styled.div`
-  font-size: 16px;
 `
 
 const PayrollTotal = styled.div`
